requirements.txt
psycopg2
asgiref
Django
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
djangorestframework
django-filter
slug
pytest
pytest-django